+++ src/ScottPlot5/ScottPlot5/Rendering/RenderManager.cs
@@ -87,6 +87,14 @@ public class RenderManager(Plot plot)
     /// </summary>
     public int RenderCount { get; private set; } = 0;
 
+    /// <summary>
+    /// Remove all render actions of the given type
+    /// </summary>
+    public void Remove<T>() where T : IRenderAction
+    {
+        RenderActions.RemoveAll(x => x is T);
+    }
+
     public void Render(SKCanvas canvas, PixelRect rect)
     {
         if (EnableRendering == false)